backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
alembic>=1.14.0
pyjwt>=2.10.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.10.0
email-validator>=2.2.0
chromadb>=0.5.23
llama-index-core>=0.12.52
llama-index-llms-openai>=0.4.7
llama-index-llms-ollama>=0.6.2
llama-index-embeddings-huggingface>=0.2.2
llama-index-embeddings-openai>=0.4.0
llama-index-embeddings-ollama>=0.6.0
llama-index-readers-file>=0.1.23
llama-index-vector-stores-chroma>=0.4.2
sentence-transformers>=3.3.0
openai>=1.58.0
ollama>=0.4.0
genanki>=0.13.1
toml>=0.10.2
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0